==> R/04_ScheduleHarmonization.R <==
library(taskscheduleR)
taskscheduler_delete("COVerAGE-DB-harmonization")
taskscheduler_create(taskname = "COVerAGE-DB-harmonization", 
                     rscript =  here::here("R","04_harmonize_age_final.R"), 
                     schedule = "ONCE",
                     days = c("WED"),
                     starttime = "12:45")
